Feature/365 azure blob storage (#603)
* Squashed Azure storage commits

* Add worker blobstorage class

* Update worker deps

* Fixes and clean up (working)

* tidy

* test vs the last release

* Added helper script to update python req within docker container

* Fix for CVE-2022-21712
