+++ conf.ini
@@ -20,8 +20,9 @@ SECRET_KEY=OmuudYrSFVxcIVIWf6YlYdkP6NXApP
 TOKEN_SIGINING_KEY=JsVzvtWw2EwksaYCZsMmd2zmm
 TOKEN_REFRESH_ROTATE = True
 #PORTFOLIO_PARQUET_STORAGE = True
-
 #TOKEN_REFRESH_LIFETIME = minutes=0, hours=0, days=0, weeks=0
+
+# --- example S3 storage settings ---  #
 #STORAGE_TYPE = S3
 #AWS_BUCKET_NAME=example-bucket
 #AWS_S3_ENDPOINT_URL=http://localhost:4572
@@ -29,6 +30,14 @@ TOKEN_REFRESH_ROTATE = True
 #AWS_SECRET_ACCESS_KEY=bar
 #AWS_QUERYSTRING_EXPIRE=180
 #AWS_QUERYSTRING_AUTH=True
+
+# --- example Azure storage Settings --- #
+#STORAGE_TYPE=AZURE
+#AZURE_ACCOUNT_NAME=<storage-account-name>
+#AZURE_ACCOUNT_KEY=<storage-account-key>
+#AZURE_CONTAINER=<storage-container-name>
+#AZURE_LOCATION=server
+#AZURE_SHARED_CONTAINER=True
 
 # JWT authentication
 # Set to 'keycloak' to enable keycloak integration for authentication and user management
@@ -40,6 +49,7 @@ TOKEN_REFRESH_ROTATE = True
 
 # Levels that require admin account to use
 ANALYSIS_PRIORITY_ADMIN_LEVELS = [8, 9, 10]
+
 
 [worker]
 DISABLE_WORKER_REG = False
@@ -55,12 +65,23 @@ INPUT_GENERATION_CHUNK_SIZE = 5000
 
 #   --- Example settings ---  #
 #BASE_RUN_DIR = /home/worker/run
+
+
+# --- example S3 storage settings ---  #
 #STORAGE_TYPE = S3
 #AWS_BUCKET_NAME=example-bucket
 #AWS_ACCESS_KEY_ID=<worker-key-id>
 #AWS_SECRET_ACCESS_KEY=<worker-access-key>
 #AWS_QUERYSTRING_EXPIRE=180
 #AWS_QUERYSTRING_AUTH=True
+
+# --- example Azure storage settings --- #
+#STORAGE_TYPE=AZURE
+#AZURE_ACCOUNT_NAME=<storage-account-name>
+#AZURE_ACCOUNT_KEY=<storage-account-key>
+#AZURE_CONTAINER=<storage-container-name>
+#AZURE_LOCATION=worker
+#AZURE_SHARED_CONTAINER=True
 
 [celery]
