+++ requirements-worker.txt
@@ -17,6 +17,12 @@ attrs==21.2.0
     #   fiona
     #   jsonschema
     #   pytest
+azure-core==1.22.1
+    # via
+    #   -r requirements-worker.in
+    #   azure-storage-blob
+azure-storage-blob==12.9.0
+    # via -r requirements-worker.in
 billiard==3.6.4.0
     # via
     #   -r requirements-worker.in
@@ -34,8 +40,11 @@ celery==5.1.2
 certifi==2021.5.30
     # via
     #   fiona
+    #   msrest
     #   pyproj
     #   requests
+cffi==1.15.0
+    # via cryptography
 chainmap==1.0.3
     # via oasislmf
 chardet==4.0.0
@@ -69,6 +78,8 @@ cookiecutter==1.7.3
     # via oasislmf
 cramjam==2.5.0
     # via fastparquet
+cryptography==36.0.1
+    # via azure-storage-blob
 fasteners==0.16.3
     # via -r requirements-worker.in
 fastparquet==0.8.0
@@ -87,6 +98,8 @@ idna==3.2
     # via requests
 iniconfig==1.1.1
     # via pytest
+isodate==0.6.1
+    # via msrest
 jinja2==3.0.1
     # via
     #   cookiecutter
@@ -109,6 +122,8 @@ markupsafe==2.0.1
     # via jinja2
 msgpack==1.0.2
     # via oasislmf
+msrest==0.6.21
+    # via azure-storage-blob
 munch==2.5.0
     # via fiona
 numba==0.53.1
@@ -120,13 +135,14 @@ numpy==1.20.3
     #   fastparquet
     #   numba
     #   numexpr
-    #   oasislmf
     #   pandas
     #   pyarrow
     #   scikit-learn
     #   scipy
-oasislmf[extra]==1.23.1
+oasislmf[extra]==1.25.1
     # via -r requirements-worker.in
+oauthlib==3.2.0
+    # via requests-oauthlib
 packaging==21.0
     # via pytest
 pandas==1.3.1
@@ -148,6 +164,8 @@ py==1.10.0
     # via pytest
 pyarrow==4.0.1
     # via oasislmf
+pycparser==2.21
+    # via cffi
 pymysql==1.0.2
     # via -r requirements-worker.in
 pyparsing==2.4.7
@@ -174,9 +192,14 @@ redis==3.5.3
     # via -r requirements-worker.in
 requests==2.26.0
     # via
+    #   azure-core
     #   cookiecutter
+    #   msrest
     #   oasislmf
+    #   requests-oauthlib
     #   requests-toolbelt
+requests-oauthlib==1.3.1
+    # via msrest
 requests-toolbelt==0.9.1
     # via oasislmf
 rtree==0.9.7
@@ -196,10 +219,12 @@ shutilwhich==1.1.0
 six==1.16.0
     # via
     #   anytree
+    #   azure-core
     #   click-repl
     #   cookiecutter
     #   fasteners
     #   fiona
+    #   isodate
     #   jsonschema
     #   munch
     #   pathlib2
