+++ src/model_execution_worker/tasks.py
@@ -30,7 +30,11 @@ from pathlib2 import Path
 from ..conf import celeryconf as celery_conf
 from ..conf.iniconf import settings
 from ..common.data import STORED_FILENAME, ORIGINAL_FILENAME
-from .orig_storage_manager import StorageSelector
+
+#from .storage_manager import StorageSelector
+from .orig_storage_manager import BaseStorageConnector
+from .backends.aws_storage import AwsObjectStore
+from .backends.azure_storage import AzureObjectStore
 
 '''
 Celery task wrapper for Oasis ktools calculation.
@@ -42,8 +46,18 @@ RUNNING_TASK_STATUS = OASIS_TASK_STATUS["running"]["id"]
 app = Celery()
 app.config_from_object(celery_conf)
 logging.info("Started worker")
-filestore = StorageSelector(settings)
 debug_worker = settings.getboolean('worker', 'DEBUG', fallback=False)
+
+# Set storage manager
+selected_storage = settings.get('worker', 'STORAGE_TYPE', fallback="").lower()
+if selected_storage in ['local-fs', 'shared-fs']:
+    filestore = BaseStorageConnector(settings)
+elif selected_storage in ['aws-s3', 'aws', 's3']:
+    filestore = AwsObjectStore(settings)
+elif selected_storage in ['azure']:
+    filestore = AzureObjectStore(settings)
+else:
+    raise OasisException('Invalid value for STORAGE_TYPE: {}'.format(selected_storage))
 
 
 class TemporaryDir(object):
