+++ src/server/oasisapi/analyses/tasks.py
@@ -13,6 +13,8 @@ from urllib.request import urlopen
 from ....conf import celeryconf as celery_conf
 
 from botocore.exceptions import ClientError as S3_ClientError
+from azure.core.exceptions import ResourceNotFoundError as Blob_ResourceNotFoundError
+from azure.storage.blob import BlobLeaseClient
 from celery import Task
 from celery import signals
 from celery.result import AsyncResult
@@ -28,13 +30,15 @@ from django.db.models import When, Case, Value, F
 from django.http import HttpRequest
 from django.utils import timezone
 
-from src.conf.iniconf import settings
 from src.server.oasisapi.files.models import RelatedFile
 from src.server.oasisapi.files.views import handle_json_data
 from src.server.oasisapi.schemas.serializers import ModelParametersSerializer
+from src.server.oasisapi.files.upload import wait_for_blob_copy
+
 from .models import AnalysisTaskStatus
 from .task_controller import get_analysis_task_controller
 from ..celery_app import celery_app
+
 
 logger = get_task_logger(__name__)
 
@@ -61,6 +65,17 @@ def is_in_bucket(object_key):
                 return False
             else:
                 raise e
+
+def is_in_container(object_key):
+    if not hasattr(default_storage, 'azure_container'):
+        return False
+    else:
+        try:
+            blob = default_storage.client.get_blob_client(object_key)
+            blob.get_blob_properties()
+            return True
+        except Blob_ResourceNotFoundError:
+            return False
 
 
 def store_file(reference, content_type, creator, required=True, filename=None):
@@ -122,7 +137,35 @@ def store_file(reference, content_type, creator, required=True, filename=None):
         stored_file.obj.wait_until_exists()
         return new_related_file
 
+    # Issue Azure object Copy
+    if is_in_container(reference):
+        new_filename = filename if filename else ref
+        fname = default_storage._get_valid_path(new_filename)
+        source_blob = default_storage.client.get_blob_client(reference)
+        dest_blob = default_storage.client.get_blob_client(fname)
+
+        try:
+            lease = BlobLeaseClient(source_blob)
+            lease.acquire()
+            dest_blob.start_copy_from_url(source_blob.url)
+            wait_for_blob_copy(dest_blob)
+            lease.break_lease()
+        except Exception as e:
+            # copy failed, break file lease and re-raise
+            lease.break_lease()
+            raise e
+
+        stored_blob = default_storage.open(os.path.basename(fname))
+        new_related_file = RelatedFile.objects.create(
+            file=File(stored_blob, name=fname),
+            filename=fname,
+            content_type=content_type,
+            creator=creator,
+            store_as_filename=True)
+        return new_related_file
+
     try:
+        # Copy via shared FS
         ref = str(os.path.basename(reference))
         fname = filename if filename else ref
         return RelatedFile.objects.create(
