+++ src/server/oasisapi/portfolios/serializers.py
@@ -9,11 +9,14 @@ from django.core.files.storage import default_storage
 from drf_yasg.utils import swagger_serializer_method
 from rest_framework import serializers
 from rest_framework.exceptions import ValidationError
+from azure.core.exceptions import ResourceNotFoundError as Blob_ResourceNotFoundError
+from azure.storage.blob import BlobLeaseClient
 
 from .models import Portfolio
 from ..analyses.serializers import AnalysisSerializer
 from ..files.models import RelatedFile
 from ..files.models import file_storage_link
+from ..files.upload import wait_for_blob_copy
 from ..permissions.group_auth import validate_and_update_groups, validate_user_is_owner
 from ..schemas.serializers import (
     LocFileSerializer,
@@ -218,9 +221,8 @@ class PortfolioStorageSerializer(serializers.ModelSerializer):
         return file_storage_link(instance.reinsurance_scope_file, True)
 
     def is_in_storage(self, value):
-        if not hasattr(default_storage, 'bucket'):
-            return default_storage.exists(value)
-        else:
+        # Check AWS storage
+        if hasattr(default_storage, 'bucket'):
             try:
                 default_storage.bucket.Object(value).load()
                 return True
@@ -229,6 +231,17 @@ class PortfolioStorageSerializer(serializers.ModelSerializer):
                     return False
                 else:
                     raise e
+        # Check Azure Blob storage
+        elif hasattr(default_storage, 'azure_container'):
+            try:
+                blob = default_storage.client.get_blob_client(value)
+                blob.get_blob_properties()
+                return True
+            except Blob_ResourceNotFoundError:
+                return False
+        else:
+            return default_storage.exists(value)
+
 
     def user_belongs_to_file_group(self, value):
         file = RelatedFile.objects.filter(file=value)
@@ -308,6 +321,36 @@ class PortfolioStorageSerializer(serializers.ModelSerializer):
                 stored_file = default_storage.open(new_related_file.file.name)
                 stored_file.obj.copy({"Bucket": bucket.name, "Key": validated_data[field]})
                 stored_file.obj.wait_until_exists()
+
+            elif hasattr(default_storage, 'azure_container'):
+                # https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-copy?tabs=python
+                fname = path.basename(validated_data[field])
+                new_file_name = default_storage.get_alternative_name(validated_data[field], '')
+                new_blobname = path.basename(new_file_name)
+
+                # Copies a blob asynchronously.
+                source_blob = default_storage.client.get_blob_client(validated_data[field])
+                dest_blob = default_storage.client.get_blob_client(new_file_name)
+
+                try:
+                    lease = BlobLeaseClient(source_blob)
+                    lease.acquire()
+                    dest_blob.start_copy_from_url(source_blob.url)
+                    wait_for_blob_copy(dest_blob)
+                    lease.break_lease()
+                except Exception as e:
+                    # copy failed, break file lease and re-raise
+                    lease.break_lease()
+                    raise e
+
+                stored_blob = default_storage.open(new_blobname)
+                new_related_file = RelatedFile.objects.create(
+                    file=File(stored_blob, name=new_blobname),
+                    filename=fname,
+                    content_type=content_type,
+                    creator=self.context['request'].user,
+                    store_as_filename=True,
+                )
 
             # Shared-fs
             else:
